refactor: remove getMechRequestsCount function and update MechAgentMetrics to use mech metrics directly from the API
