+++ components/AgentEconomies/MechPage/MechAgentMetrics.jsx
@@ -1,4 +1,4 @@
-import { getMechMetrics, getMechRequestsCount } from 'common-util/api';
+import { getMechMetrics } from 'common-util/api';
 import { getMechTxs } from 'common-util/api/dune';
 import SectionWrapper from 'components/Layout/SectionWrapper';
 import { Card } from 'components/ui/card';
@@ -9,9 +9,8 @@ import { useMemo } from 'react';
 import useSWR from 'swr';
 
 const fetchMetrics = async () => {
-  const [dailyActiveAgents, totalTxs, result] = await Promise.allSettled([
+  const [mechMetrics, result] = await Promise.allSettled([
     getMechMetrics(),
-    getMechRequestsCount(),
     getMechTxs(),
   ]);
 
@@ -19,8 +18,13 @@ const fetchMetrics = async () => {
 
   return {
     dailyActiveAgents:
-      dailyActiveAgents.status === 'fulfilled' ? dailyActiveAgents.value : null,
-    totalTxs: totalTxs.status === 'fulfilled' ? totalTxs.value?.total : null,
+      mechMetrics.status === 'fulfilled'
+        ? mechMetrics.value?.dailyActiveAgents
+        : null,
+    totalTxs:
+      mechMetrics.status === 'fulfilled'
+        ? mechMetrics.value?.totalRequests?.total
+        : null,
     predictTxs: mechTxs?.predictTxs || null,
     contributeTxs: mechTxs?.contributeTxs || null,
     governatooorrTxs: mechTxs?.governatooorrTxs || null,
@@ -142,9 +146,13 @@ export const MechAgentMetrics = () => {
         <p className="text-xl text-slate-700 mb-8 mx-auto mt-12">
           The Olas Mech agent economy is in demand as ever, resulting in more
           than{' '}
-          <Link className="font-bold" href="/data#mech-requests" hideArrow>
-            {metrics?.totalTxs?.toLocaleString()}
-          </Link>{' '}
+          {typeof metrics?.totalTxs === 'number' ? (
+            <Link className="font-bold" href="/data#mech-requests" hideArrow>
+              {metrics.totalTxs.toLocaleString()}
+            </Link>
+          ) : (
+            <span className="font-bold">--</span>
+          )}{' '}
           requests from other AI agent economies.
         </p>
       </div>
