+++ pages/api/mech-metrics.js
@@ -1,6 +1,13 @@
 import { calculate7DayAverage } from 'common-util/calculate7DayAverage';
-import { REGISTRY_GRAPH_CLIENTS } from 'common-util/graphql/client';
-import { dailyMechAgentPerformancesQuery } from 'common-util/graphql/queries';
+import {
+  ATA_GRAPH_CLIENTS,
+  REGISTRY_GRAPH_CLIENTS,
+} from 'common-util/graphql/client';
+import {
+  dailyMechAgentPerformancesQuery,
+  mechGlobalsTotalRequestsQuery,
+  mechMarketplaceTotalRequestsQuery,
+} from 'common-util/graphql/queries';
 import { getMidnightUtcTimestampDaysAgo } from 'common-util/time';
 
 const CACHE_DURATION_SECONDS = 12 * 60 * 60; // 12 hours
@@ -42,6 +49,48 @@ const fetchDailyAgentPerformance = async () => {
   }
 };
 
+const fetchMechRequestsFromSubgraphs = async () => {
+  try {
+    const [legacyMechResult, marketplaceGnosisResult, marketplaceBaseResult] =
+      await Promise.allSettled([
+        ATA_GRAPH_CLIENTS.legacyMech.request(mechGlobalsTotalRequestsQuery),
+        ATA_GRAPH_CLIENTS.gnosis.request(mechMarketplaceTotalRequestsQuery),
+        ATA_GRAPH_CLIENTS.base.request(mechMarketplaceTotalRequestsQuery),
+      ]);
+
+    const legacyMechTotalRequests = (() => {
+      if (legacyMechResult.status !== 'fulfilled') return 0;
+      const response = legacyMechResult.value;
+      return Number(response?.global?.totalRequests ?? 0);
+    })();
+
+    const marketplaceGnosisTotalRequests = (() => {
+      if (marketplaceGnosisResult.status !== 'fulfilled') return 0;
+      const response = marketplaceGnosisResult.value;
+      return Number(response?.global?.totalRequests ?? 0);
+    })();
+
+    const marketplaceBaseTotalRequests = (() => {
+      if (marketplaceBaseResult.status !== 'fulfilled') return 0;
+      const response = marketplaceBaseResult.value;
+      return Number(response?.global?.totalRequests ?? 0);
+    })();
+
+    return {
+      mech: legacyMechTotalRequests,
+      marketplace:
+        marketplaceGnosisTotalRequests + marketplaceBaseTotalRequests,
+      total:
+        legacyMechTotalRequests +
+        marketplaceGnosisTotalRequests +
+        marketplaceBaseTotalRequests,
+    };
+  } catch (error) {
+    console.error('Error fetching mech requests from subgraphs:', error);
+    return null;
+  }
+};
+
 export default async function handler(req, res) {
   if (req.method !== 'GET') {
     return res.status(405).json({ message: 'Method not allowed' });
@@ -58,18 +107,20 @@ export default async function handler(req, res) {
   );
 
   try {
-    const latestMetrics = await fetchDailyAgentPerformance();
-    if (latestMetrics) {
-      return res.status(200).json(latestMetrics);
+    const [dailyActiveAgents, requests] = await Promise.all([
+      fetchDailyAgentPerformance(),
+      fetchMechRequestsFromSubgraphs(),
+    ]);
+
+    if (dailyActiveAgents !== null && requests !== null) {
+      return res
+        .status(200)
+        .json({ dailyActiveAgents, totalRequests: requests });
     }
 
-    return res
-      .status(404)
-      .json({ error: 'Error fetching Daily Active Agents for mechs.' });
+    return res.status(404).json({ error: 'Error fetching mech metrics.' });
   } catch (error) {
     console.error('Error in handler:', error);
-    return res
-      .status(500)
-      .json({ error: 'Failed to fetch agent performance values.' });
+    return res.status(500).json({ error: 'Failed to fetch mech metrics.' });
   }
 }
